Let's use a real database.
- Created skeleton for SQLite operations.
- Main RecyclerView now loads tasks from DB rather than hardcoded ones
- LTRecyclerView should abstract some of the stupidity out.
- Randomly chose a FAB library, let's see how it is.

NOTE: Database stuff might change since I haven't actually added any items
to see if the logic holds up yet.

+++ app/src/main/java/com/kasra/lifetracker/models/TaskEvent.java
@@ -5,7 +5,7 @@ import android.location.Location;
 import java.util.Date;
 
 public class TaskEvent {
-    public int _id;
+    public long _id;
     public Date time;
     public Location location;
 }
